requirements.txt
asgiref==3.3.4
cffi==1.14.5
Django==3.2
Pillow==8.2.0
pycparser==2.20
pylokit==0.8.1
pytesseract==0.3.7
pytz==2021.1
six==1.15.0
sqlparse==0.4.1
typing-extensions==3.7.4.3
